no more PM abe Japan
A guys of modern Japanese people needs the SDF in order to preserve the descendants of their descendants. The idea is totally different from the left wing who dislikes the army by the defeat, now the old people. So the motivation of the Constitution clearly decreases. The SDF is supported by us.